feat: refactor control label components
