+++ src/components/Controls/ControlLabel/ControlLabel.style.ts
@@ -3,8 +3,13 @@ import { css } from '@emotion/react';
 import type { Theme } from 'theme';
 import type { SemanticTypographyKey } from 'theme/tokens/semantic/typography';
 
-import type { LabelConfig } from '../Controls.types';
 import { generateStylesFromTokens } from 'components/Typography/utils';
+import type { LabelConfig } from '../Controls.types';
+
+const labelSize: Record<string, SemanticTypographyKey> = {
+  normal: 'normal.body02',
+  large: 'normal.title01',
+};
 
 export const labelContainerStyles = () => (theme: Theme) => {
   return css`
@@ -12,11 +17,6 @@ export const labelContainerStyles = () => (theme: Theme) => {
     flex-direction: column;
     gap: ${theme.dimension.spacing.get('2xs')};
   `;
-};
-
-const labelSize: Record<string, SemanticTypographyKey> = {
-  normal: 'normal.body02',
-  large: 'normal.headline04',
 };
 
 export const labelStyles =
@@ -34,5 +34,6 @@ export const helpTextStyles =
     return css`
       ${generateStylesFromTokens(theme.tokens.typography.get('normal.body03'))};
       color: ${theme.tokens.colors.get('textColor.default.secondary')};
+      padding-top: ${theme.globals.spacing.get('2')};
     `;
   };
