+++ src/components/Controls/ControlLabel/index.ts
@@ -1,1 +1,2 @@
-export { default } from './ControlLabel';
+export { default as ControlHelpText } from './ControlHelpText';
+export { default as ControlLabelText } from './ControlLabelText';
